# app/docs/swagger.yaml
definitions:
...
  domain.UploadedFile:
    properties:
      createTime:
        type: string
      isDir:
        type: boolean
      mime:
        type: string
      name:
        type: string
      size:
        type: integer
      updateTime:
        type: string
      url:
        type: string
    type: object
...
host: localhost:9123
info:
  contact: {}
  description: FireBoom-server
  title: FireBoom-server Swagger API
  version: "1.0"
paths:
...
  /api/v1/file/upload:
    post:
      consumes:
      - application/json
      description: 添加保存文件
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - file
...
  /api/v1/hook/script:
    post:
      consumes:
      - application/json
      description: 保存钩子内容
      parameters:
      - description: 钩子路径
        in: formData
        name: path
        required: true
        type: string
      - description: 钩子类型
        in: formData
        name: switch
        required: true
        type: boolean
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - hook
...
  /api/v1/s3Upload/detail:
    post:
      consumes:
      - application/json
      description: 文件详情
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: 文件名称
        in: formData
        name: fileName
        required: true
        type: string
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/domain.UploadedFile'
        "400":
          description: 查询失败
      tags:
      - s3Upload
  /api/v1/s3Upload/download:
    get:
      consumes:
      - application/json
      description: 下载
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: 文件名称
        in: formData
        name: fileName
        required: true
        type: string
      responses:
        "200":
          description: 下载成功
        "400":
          description: 下载失败
      tags:
      - s3Upload
  /api/v1/s3Upload/list:
    get:
      consumes:
      - application/json
      description: 查询文件列表
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: 前缀
        in: formData
        name: filePrefix
        required: true
        type: string
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.UploadedFile'
            type: array
        "400":
          description: 查询失败
      tags:
      - s3Upload
  /api/v1/s3Upload/remove:
    post:
      consumes:
      - application/json
      description: 删除
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: 文件名称
        in: formData
        name: fileName
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - s3Upload
  /api/v1/s3Upload/rename:
    put:
      consumes:
      - application/json
      description: 重命名
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: 旧名称
        in: formData
        name: oldName
        required: true
        type: string
      - description: 新名词
        in: formData
        name: newName
        required: true
        type: string
      responses:
        "200":
          description: 重命名成功
        "400":
          description: 重命名失败
      tags:
      - s3Upload
  /api/v1/s3Upload/upload:
    post:
      consumes:
      - application/json
      description: oss上传
      parameters:
      - description: 储存桶id
        in: formData
        name: bucketID
        required: true
        type: integer
      - description: oss路径
        in: formData
        name: path
        required: true
        type: string
      responses:
        "200":
          description: 上传成功
        "400":
          description: 上传失败
      tags:
      - s3Upload
  /api/v1/script:
    get:
      consumes:
      - application/json
      description: 根据脚本名获取脚本
      parameters:
      - description: 名称
        in: formData
        name: scriptName
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - script
    post:
      consumes:
      - application/json
      description: 添加脚本
      parameters:
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 内容
        in: formData
        name: content
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - script
...
  /api/v1/wdg/close:
    get:
      consumes:
      - application/json
      description: 停止
      responses:
        "200":
          description: 停止成功
        "400":
          description: 停止失败
      tags:
      - wdg
...
  /api/v1/wdg/reStart:
    get:
      consumes:
      - application/json
      description: 重启
      responses:
        "200":
          description: 重启成功
        "400":
          description: 重启失败
      tags:
      - wdg
  /api/v1/wdg/start:
    get:
      consumes:
      - application/json
      description: 启动
      responses:
        "200":
          description: 启动成功
        "400":
          description: 启动失败
      tags:
      - wdg
...
produces:
- application/json
swagger: "2.0"